modularize code and merge Joeys updates
