+++ Scratch/communitytest.py
