+++ Scratch/locationtest2.py
