+++ Scratch/simplestlocation.py
