+++ assets/style.css
@@ -1,5 +1,5 @@
-.custom-dropdown .Select--multi .Select-value {
+/* .custom-dropdown .Select--multi .Select-value {
     background-color: #fcba03;
     border-color: #e3a700;
     color: #fff;
-  }
+  } */
